Extract LLM service into packages/team-llm internal package
Spin out models/llm core (LLMService, providers, config, retry, exceptions)
into a standalone team-llm package under packages/ for internal reuse.

- Add packages/team-llm with pyproject.toml, src layout, and tests
- Add SETUP.md with GitHub repo, uv setup, and install instructions
- Wire monorepo via editable path dependency in pyproject.toml
- Keep models.llm as thin compatibility re-exports for existing imports
