+++ packages/team-llm/src/team_llm/config/__init__.py
@@ -1,0 +1,1 @@
+"""Model configuration."""
